Bugfix in repositories

+++ src/main/java/edu/lb/spring_networktechnologies/infrastructure/repositores/LoanRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LoanRepository extends CrudRepository<Loan, Integer> {
+public interface LoanRepository extends CrudRepository<Loan, Long> {
 }
